♻️ refactor: refactor the DragUpload component
