+++ src/app/(main)/chat/(workspace)/@conversation/features/ChatInput/Desktop/Footer/index.tsx
@@ -7,6 +7,7 @@ import { memo, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Center, Flexbox } from 'react-layout-kit';
 
+import DragUpload from '@/components/DragUpload';
 import StopLoadingIcon from '@/components/StopLoading';
 import SaveTopic from '@/features/ChatInput/Topic';
 import { useSendMessage } from '@/features/ChatInput/useSend';
@@ -19,7 +20,6 @@ import { useUserStore } from '@/store/user';
 import { modelProviderSelectors, preferenceSelectors } from '@/store/user/selectors';
 import { isMacOS } from '@/utils/platform';
 
-import DragUpload from './DragUpload';
 import { LocalFiles } from './LocalFiles';
 import SendMore from './SendMore';
 
@@ -77,10 +77,26 @@ const Footer = memo<FooterProps>(({ setExpand }) => {
 
   const model = useAgentStore(agentSelectors.currentAgentModel);
 
+  const enabledFiles = useUserStore(modelProviderSelectors.isModelEnabledFiles(model));
+
   const [useCmdEnterToSend, canUpload] = useUserStore((s) => [
     preferenceSelectors.useCmdEnterToSend(s),
     modelProviderSelectors.isModelEnabledUpload(model)(s),
   ]);
+
+  const uploadFile = useFileStore((s) => s.uploadFile);
+
+  const uploadImages = async (fileList: FileList | undefined) => {
+    if (!fileList || fileList.length === 0) return;
+
+    const pools = Array.from(fileList).map(async (file) => {
+      // skip none-file items
+      if (!file.type.startsWith('image') && !enabledFiles) return;
+      await uploadFile(file);
+    });
+
+    await Promise.all(pools);
+  };
 
   const sendMessage = useSendMessage();
 
@@ -128,7 +144,7 @@ const Footer = memo<FooterProps>(({ setExpand }) => {
       <Flexbox align={'center'} gap={8} horizontal style={{ overflow: 'hidden' }}>
         {canUpload && (
           <>
-            <DragUpload />
+            <DragUpload enabledFiles={enabledFiles} onUploadFiles={uploadImages} />
             <LocalFiles />
           </>
         )}
