+++ src/locales/default/chat.ts
@@ -132,9 +132,5 @@ export default {
     actionFiletip: '上传文件',
     actionTooltip: '上传图片',
     disabled: '当前模型不支持视觉识别和文件分析，请切换模型后使用',
-    dragDesc: '拖拽文件到这里，支持上传多个图片。按住 Shift 直接发送图片',
-    dragFileDesc: '拖拽图片和文件到这里，支持上传多个图片和文件。按住 Shift 直接发送图片或文件',
-    dragFileTitle: '上传文件',
-    dragTitle: '上传图片',
   },
 };
